requirements.txt
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
coverage==6.4.2
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
imageio==2.20.0
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
moviepy==1.0.3
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
proglog==0.1.10
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-html==3.1.1
pytest-metadata==2.0.2
requests==2.28.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.11
